use new buttons for every not started match, some smartphone improvements
