+++ src/app/table-list/match/player.model.ts
@@ -1,6 +1,7 @@
 export interface Player {
-    player_id?: number;
-    first_name?: string;
-    last_name?: string;
-    club?: string;
+  player_id?: number;
+  first_name?: string;
+  last_name?: string;
+  club?: string;
+  call_count?: number;
 }
