+++ src/app/table-list/table-list.module.ts
@@ -11,36 +11,36 @@ import {TableListComponent} from './table-list.component';
 import {ResultDialogComponent} from './tt-table/result-dialog/result-dialog.component';
 import {StartDialogComponent} from './tt-table/start-dialog/start-dialog.component';
 import {TtTableComponent} from './tt-table/tt-table.component';
-import {MatchlistComponent} from './tt-table/matchlist/matchlist.component';
 import {PlayerDialogComponent} from './tt-table/player-dialog/player-dialog.component';
-import { GameDialogComponent } from './tt-table/game-dialog/game-dialog.component';
-import { ClubNamePipe } from './club-name.pipe';
+import {GameDialogComponent} from './tt-table/game-dialog/game-dialog.component';
+import {ClubNamePipe} from './club-name.pipe';
+import {MatchItemComponent} from './tt-table/match-item/match-item.component';
 
 @NgModule({
-    declarations: [
-        TtTableComponent,
-        TableListComponent,
-        PlayerNamePipe,
-        ResultDialogComponent,
-        StartDialogComponent,
-        MatchlistComponent,
-      PlayerDialogComponent,
-      GameDialogComponent,
-      ClubNamePipe
-    ],
-    imports: [
-        SharedModule,
-        StoreModule.forFeature('tables', fromTables.reducer),
-        EffectsModule.forFeature([MatchEffects, TableListEffects]),
-        ReactiveFormsModule
-    ],
-    exports: [TableListComponent],
-    entryComponents: [
-        ResultDialogComponent,
-        StartDialogComponent,
-      PlayerDialogComponent,
-      GameDialogComponent
-    ]
+  declarations: [
+    TtTableComponent,
+    TableListComponent,
+    PlayerNamePipe,
+    ResultDialogComponent,
+    StartDialogComponent,
+    PlayerDialogComponent,
+    GameDialogComponent,
+    ClubNamePipe,
+    MatchItemComponent
+  ],
+  imports: [
+    SharedModule,
+    StoreModule.forFeature('tables', fromTables.reducer),
+    EffectsModule.forFeature([MatchEffects, TableListEffects]),
+    ReactiveFormsModule
+  ],
+  exports: [TableListComponent],
+  entryComponents: [
+    ResultDialogComponent,
+    StartDialogComponent,
+    PlayerDialogComponent,
+    GameDialogComponent
+  ]
 })
 export class TableListModule {
 }
