--- src/app/table-list/tt-table/matchlist/matchlist.component.html
@@ -1,5 +1,0 @@
-<div class="row" *ngFor="let match of matches">
-    <div class="col-12">
-        <p>{{match.players_a | playerName}} - {{match.players_b | playerName}} ({{match.result.games_won_player_a}}:{{match.result.games_won_player_b}})</p>
-    </div>
-</div>
