+++ src/app/table-list/tt-table/tt-table.component.ts
@@ -105,48 +105,12 @@ export class TtTableComponent {
     });
   }
 
-  startMatch() {
-    if (this.table.matches.length === 1) {
-      this.startMatchOnTable.emit({
-        tableId: this.table.table_id,
-        matchId: this.table.matches[0].match_id
-      });
-      return;
-    }
-    const dialogRef = this.dialog.open(StartDialogComponent, {
-      width: '400px',
-      data: this.table.matches.filter(match => match.state === 'ASSIGNED')
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (!!result) {
-        this.startMatchOnTable.emit({
-          tableId: this.table.table_id,
-          matchId: result
-        });
-      }
+  startMatch(matchId) {
+    this.startMatchOnTable.emit({
+      tableId: this.table.table_id,
+      matchId
     });
   }
-
-  callPlayer() {
-    const players = this.table.matches
-      .map(match => match.players_a.concat(match.players_b))
-      .reduce((first, second) => first.concat(second));
-    const dialogRef = this.dialog.open(PlayerDialogComponent, {
-      width: '400px',
-      data: players
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (!!result) {
-        this.updateMatchResult.emit({
-          playerIds: result,
-          matchId: this.currentMatch().match_id
-        });
-      }
-    });
-  }
-
 
   endMatch() {
     this.finishMatch.emit(this.currentMatch());
