+++ src/styles.scss
@@ -2,4 +2,5 @@
 
 @import '~bootstrap/scss/bootstrap-reboot';
 @import '~bootstrap/scss/bootstrap-grid';
+@import '~bootstrap/scss/_utilities';
 @import '~@angular/material/prebuilt-themes/deeppurple-amber.css';
